More strings cleanup. Emit a canonical winrt/base_macros.h
