+++ natvis/pch.h
@@ -12,6 +12,7 @@
 #include <Dia2.h>
 #include "base_includes.h"
 #include "base_macros.h"
+#include "base_source_location.h"
 #include "base_types.h"
 #include "base_extern.h"
 #include "base_meta.h"
